refactor(dev): tsconfig utils
